+++ Bootstrap/Attributes/UserAttributes.php
@@ -30,7 +30,11 @@ class UserAttributes extends AbstractAttributes
 
     protected function uninstallAttributes()
     {
-        $this->crudService->delete($this->tableName, 'billie_iban');
-        $this->crudService->delete($this->tableName, 'billie_bic');
+        if ($this->crudService->get($this->tableName, 'billie_iban')) {
+            $this->crudService->delete($this->tableName, 'billie_iban');
+        }
+        if ($this->crudService->get($this->tableName, 'billie_bic')) {
+            $this->crudService->delete($this->tableName, 'billie_bic');
+        }
     }
 }
